Wrote up playerInput, update functionality for multiple components and the entity class, game runs on 10 fps fml

+++ Source/Camera.cpp
@@ -5,10 +5,10 @@ float Camera::posY = 0;
 
 
 void Camera::centerAround(int x, int y, int levelWidth, int levelHeight) {
-	posX = x - Globals::SCREEN_WIDTH / 2;
-	posY = y - Globals::SCREEN_HEIGHT / 2;
+	posX = (float)(x - Globals::SCREEN_WIDTH / 2);
+	posY = (float)(y - Globals::SCREEN_HEIGHT / 2);
 	if (posX < 0) posX = 0;
 	if (posY < 0) posY = 0;
-	if (posX > levelWidth) posX = levelWidth - Globals::SCREEN_WIDTH;
-	if (posY > levelHeight) posY = levelHeight - Globals::SCREEN_HEIGHT;
+	if (posX > levelWidth) posX = (float)(levelWidth - Globals::SCREEN_WIDTH);
+	if (posY > levelHeight) posY = (float)(levelHeight - Globals::SCREEN_HEIGHT);
 }
